Make debugging easier; replace ioutil (#26)
* Change message about embed directory to be more clear

* Add argument to preserve files extracted from the input tile

* Replace uses of deprecated ioutil with appropriate io/os fuctions

* Update tests to match new argument and embed info message

* Fix whitespace for help message

* Only print tile extraction directory when --preserve-extracted is set

* Split the checks for existing file system and missing file system

* Added test for --preserve-extracted flag

+++ main.go
@@ -2,7 +2,6 @@ package main
 
 import (
 	"fmt"
-	"io/ioutil"
 	"log"
 	"os"
 
@@ -14,18 +13,20 @@ import (
 const usageText = `winfs-injector injects the Windows 2016 root file system into the Windows 2016 Runtime Tile.
 
 Usage: winfs-injector
-  --input-tile, -i   path to input tile (example: /path/to/input.pivotal)
-  --output-tile, -o  path to output tile (example: /path/to/output.pivotal)
-  --registry, -r     path to docker registry (example: /path/to/registry, default: "https://registry.hub.docker.com")
-  --help, -h         prints this usage information
+  --input-tile, -i		path to input tile (example: /path/to/input.pivotal)
+  --output-tile, -o		path to output tile (example: /path/to/output.pivotal)
+  --preserve-extracted, -p	preserve the files created during the tile extraction process (useful for debugging)
+  --registry, -r		path to docker registry (example: /path/to/registry, default: "https://registry.hub.docker.com")
+  --help, -h			prints this usage information
 `
 
 func main() {
 	var arguments struct {
-		InputTile  string `short:"i" long:"input-tile"`
-		OutputTile string `short:"o" long:"output-tile"`
-		Registry   string `short:"r" long:"registry" default:"https://registry.hub.docker.com"`
-		Help       bool   `short:"h" long:"help"`
+		InputTile         string `short:"i" long:"input-tile"`
+		OutputTile        string `short:"o" long:"output-tile"`
+		PreserveExtracted bool   `short:"p" long:"preserve-extracted"`
+		Registry          string `short:"r" long:"registry" default:"https://registry.hub.docker.com"`
+		Help              bool   `short:"h" long:"help"`
 	}
 
 	_, err := jhanda.Parse(&arguments, os.Args[1:])
@@ -42,11 +43,16 @@ func main() {
 	var zipper = tile.NewZipper()
 	var releaseCreator = winfsinjector.ReleaseCreator{}
 
-	wd, err := ioutil.TempDir("", "")
+	wd, err := os.MkdirTemp("", "")
 	if err != nil {
 		log.Fatal(err)
 	}
-	defer os.RemoveAll(wd)
+
+	if arguments.PreserveExtracted {
+		fmt.Fprintf(os.Stdout, "tile extraction directory: %s\n", wd)
+	} else {
+		defer os.RemoveAll(wd)
+	}
 
 	app := winfsinjector.NewApplication(releaseCreator, tileInjector, zipper)
